Version Packages (#302)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## ai@2.1.17

### Patch Changes

-   3ed65bf: Remove dependency on node crypto API